+++ src/components/Board/index.module.less
@@ -1,13 +1,6 @@
-//.boardPanel {
-//  padding: 24px 32px;
-//  background: var(--td-bg-color-container);
-//  border-radius: 3px;
-//  width: 100%;
-//  display: flex;
-//  flex-direction: column;
-//  height: 168px;
-//  position: relative;
-//}
+.panel {
+  height: 160px
+}
 
 .boardPanelDark {
   background: var(--td-brand-color);
@@ -20,22 +13,6 @@
 }
 
 .boardTitle {}
-
-//.boardTitle {
-//  display: flex;
-//  justify-content: space-between;
-//  font-size: 14px;
-//  line-height: 22px;
-//  color: var(--td-text-color-secondary);
-//  margin-bottom: 8px;
-//}
-//
-//.boardContainer {
-//  flex: 1;
-//  display: flex;
-//  flex-direction: column;
-//  justify-content: space-between;
-//}
 
 .boardItem {
   display: flex;
