--- src/pages/Dashboard/Base/index.module.less
@@ -1,3 +1,0 @@
-.dashboardPanel {
-  margin: 24px;
-}
