--- src/pages/Dashboard/Detail/index.module.less
@@ -1,3 +1,0 @@
-.dashboardDetail {
-  margin: 24px;
-}
