+++ src/pages/Dashboard/Detail/index.tsx
@@ -1,13 +1,13 @@
 import React, { memo } from 'react';
+import PageBox from 'components/PageBox';
 import MonthPurchase from './components/MonthPurchase';
 import PurchaseTrend from './components/PurchaseTrend';
 import PurchaseSatisfaction from './components/Satisfaction';
-import Style from './index.module.less';
 
 export default memo(() => (
-  <div className={Style.dashboardDetail}>
+  <PageBox>
     <MonthPurchase />
     <PurchaseTrend />
     <PurchaseSatisfaction />
-  </div>
+  </PageBox>
 ));
