+++ src/pages/Detail/Advanced/components/Progress.tsx
@@ -7,7 +7,7 @@ import Style from '../index.module.less';
 const { StepItem } = Steps;
 
 const Progress = () => (
-  <Card borded={false} title='发票进度' className={Style.cardBox}>
+  <Card title='发票进度' className={Style.cardBox}>
     <Steps current={stepCurrent}>
       {dataStep.map((item) => (
         <StepItem key={item.id} title={item.name} content={item.detail} />
