+++ src/pages/Detail/Deploy/components/BottomTable.module.less
@@ -1,0 +1,10 @@
+.operationLink {
+  color: var(--td-brand-color);
+  text-decoration: none;
+  margin-right: 24px;
+  cursor: pointer;
+  transition: color 0.2s cubic-bezier(0.38, 0, 0.24, 1);
+  &:last-child {
+    margin-right: 0;
+  }
+}
