+++ src/pages/Detail/Deploy/components/ManagementPopup.module.less
@@ -1,0 +1,30 @@
+.popupBox {
+  max-height: 620px;
+}
+
+.popupItem {
+  padding: 12px 0;
+  display: flex;
+  height: 46px;
+  align-items: center;
+  box-sizing: border-box;
+  > h1 {
+    margin: 0;
+    width: 84px;
+    font-size: 14px;
+    color: var(--td-text-color-secondary);
+    text-align: left;
+    line-height: 22px;
+  }
+  > p {
+    margin: 0 0 0 24px;
+  }
+}
+
+.popupItem_blue {
+  color: var(--td-brand-color-8);
+}
+
+.popupItem_green {
+  color: var(--td-success-color-5);
+}
