+++ src/pages/Detail/Deploy/components/TopChart.module.less
@@ -1,0 +1,8 @@
+.panel {
+  margin-bottom: 16px;
+}
+.deployPanelLeft {
+  border-radius: 3px;
+  min-width: 230px;
+  width: 100%;
+}
