--- src/pages/Detail/Deploy/index.module.less
@@ -1,60 +1,0 @@
-.deployPanel {
-  margin: 24px;
-}
-
-.deployPanelLeft {
-  border-radius: 3;
-  min-width: 230px;
-  width: 100%;
-}
-
-.deploy-panel-left-title {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.operationLink {
-  color: var(--td-brand-color);
-  text-decoration: none;
-  margin-right: 24px;
-  cursor: pointer;
-  transition: color 0.2s cubic-bezier(0.38, 0, 0.24, 1);
-  &:last-child {
-    margin-right: 0;
-  }
-}
-
-.popupBox {
-  max-height: 620px;
-}
-
-.popupItem {
-  padding: 12px 0;
-  display: flex;
-  height: 46px;
-  align-items: center;
-  box-sizing: border-box;
-  > h1 {
-    margin: 0;
-    width: 84px;
-    font-family: PingFangSC-Regular;
-    font-size: 14px;
-    color: var(--td-text-color-secondary);
-    text-align: left;
-    line-height: 22px;
-  }
-  > p {
-    margin: 0;
-    margin-left: 24px;
-  }
-}
-
-.popupItem_blue {
-  color: var(--td-brand-color-8);
-}
-
-.popupItem_green {
-  color: var(--td-success-color-5);
-}
