+++ src/pages/List/Tree/index.module.less
@@ -1,5 +1,4 @@
 .content {
-  padding: 0;
   display: flex;
 }
